Move table components to separate files
