+++ src/components/EMS/EMS.tsx
@@ -1,134 +1,11 @@
 import React from "react";
-import { Link } from "react-router-dom";
 import { useQuery } from "urql";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faArrowUpRightFromSquare } from "@fortawesome/free-solid-svg-icons";
 
-import { formatDate } from "utils/formatters";
 import LoadingSpinner from "components/shared/LoadingSpinner";
-import Button from "components/shared/Button";
-import {
-  GetCompletedRunsDocument,
-  HistTableEntry,
-  HistTablePage,
-  PageInfo,
-} from "generated";
+import CompletedRunsTable from "./CompletedRunsTable";
+import { GetCompletedRunsDocument } from "generated";
 
 const ROWS_PER_QUERY = 15;
-
-type CompletedRunsTableProps = {
-  tableEntriesFromProps: HistTablePage["edges"];
-  setCursor: (cursor: string | null) => void;
-  pageInfo?: PageInfo | null;
-};
-const CompletedRunsTable = ({
-  tableEntriesFromProps,
-  setCursor,
-  pageInfo,
-}: CompletedRunsTableProps) => {
-  const [tableEntries, setTableEntries] = React.useState(tableEntriesFromProps);
-  const prevTableEntries = React.useRef<typeof tableEntries>();
-
-  React.useEffect(() => {
-    if (
-      prevTableEntries.current?.length &&
-      prevTableEntries.current[0]?.node?.name !== tableEntriesFromProps &&
-      tableEntriesFromProps?.length &&
-      tableEntriesFromProps[0]?.node?.name
-    ) {
-      setTableEntries([...prevTableEntries.current, ...tableEntriesFromProps]);
-    }
-  }, [tableEntriesFromProps]);
-
-  if (!tableEntries) {
-    return <LoadingSpinner className="mt-24" />;
-  }
-
-  return (
-    <div className="mx-48 my-12 text-center">
-      <div className="relative overflow-x-auto shadow-md sm:rounded-lg">
-        <table className="w-full text-sm text-left text-gray-500 dark:text-gray-400">
-          <thead className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
-            <tr>
-              <th scope="col" className="px-6 py-3">
-                Run Name
-              </th>
-              <th scope="col" className="px-6 py-3">
-                Start Date
-              </th>
-              <th scope="col" className="px-6 py-3">
-                Start Time
-              </th>
-              <th scope="col" className="px-6 py-3" />
-            </tr>
-          </thead>
-          <tbody>
-            {!tableEntries.length ? (
-              <tr className="bg-white border-b dark:bg-gray-800 dark:border-gray-700">
-                <td colSpan={4} className="px-6 py-4 text-center">
-                  No completed runs found
-                </td>
-              </tr>
-            ) : (
-              tableEntries.map((tableEntry, index) => (
-                <TableRow
-                  key={`${tableEntry?.node?.name}_${index}`}
-                  tableEntry={tableEntry?.node}
-                  isLastEntry={index === tableEntries.length - 1}
-                />
-              ))
-            )}
-          </tbody>
-        </table>
-      </div>
-      {pageInfo?.hasNextPage && (
-        <Button
-          className="px-4 py-2 mt-5"
-          onClick={() => {
-            prevTableEntries.current = tableEntries;
-            setCursor(pageInfo?.endCursor || null);
-          }}
-        >
-          Load More
-        </Button>
-      )}
-    </div>
-  );
-};
-
-type TableRowProps = {
-  tableEntry?: HistTableEntry | null;
-  isLastEntry: boolean;
-};
-const TableRow = ({ tableEntry }: TableRowProps) => {
-  if (!tableEntry) {
-    return null;
-  }
-  const runNumber = tableEntry.name?.replace(/\D/g, "") || "";
-  return (
-    <tr className="bg-white border-b dark:bg-gray-800 dark:border-gray-700">
-      <th
-        scope="row"
-        className="px-6 py-4 font-medium text-gray-900 dark:text-white whitespace-nowrap"
-      >
-        #{runNumber}
-      </th>
-      <td className="px-6 py-4">{formatDate(tableEntry.created, "date")}</td>
-      <td className="px-6 py-4">{formatDate(tableEntry.created, "time")}</td>
-      <td className="px-6 py-4 text-right">
-        <Link
-          to={`/ems/${tableEntry.name}`}
-          className="font-medium text-blue-600 dark:text-blue-500 hover:underline"
-        >
-          <FontAwesomeIcon
-            className="text-dark-blue dark:text-white"
-            icon={faArrowUpRightFromSquare}
-          />
-        </Link>
-      </td>
-    </tr>
-  );
-};
 
 const EMS = () => {
   React.useEffect(() => {
